call the permissions api on mount of the onboardings page
